requirements.txt
fastapi==0.115.12
uvicorn==0.34.2
watchfiles==1.0.5
httptools==0.6.4
python-multipart==0.0.20
SQLAlchemy==2.0.41
psycopg2-binary==2.9.10
greenlet==3.2.2
openai==1.78.1
langchain-community==0.2.6
requests==2.32.3
certifi==2025.4.26
charset-normalizer==3.4.2
idna==3.10
urllib3==2.4.0
anyio==4.9.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1 
sniffio==1.3.1
reportlab>=3.6.0
python-dotenv==1.1.0
Jinja2==3.1.6
MarkupSafe==3.0.2 
orjson==3.10.18 
pydantic==2.11.4 